skal sende med en oppsummering av samværsberegningen med til backend for journalføring
