+++ src/frontend/Felles/Kalkulator/utils.ts
@@ -1,4 +1,10 @@
-import { Samværsandel, Samværsavtale, Samværsdag, Samværsuke } from '../../App/typer/samværsavtale';
+import {
+    Samværsandel,
+    samværsandelTilVerdi,
+    Samværsavtale,
+    Samværsdag,
+    Samværsuke,
+} from '../../App/typer/samværsavtale';
 
 const samværsdag: Samværsdag = {
     andeler: [],
@@ -83,4 +89,26 @@ export const oppdaterVarighetPåSamværsavtale = (
             uker: prevState.uker.slice(0, nyVarighet),
         }));
     }
+};
+
+export const utledVisningstekst = (samværsuker: Samværsuke[]) => {
+    const summertSamvær = samværsuker
+        .flatMap((samværsuke) =>
+            Object.values(samværsuke).flatMap((samværsdag: Samværsdag) => samværsdag.andeler)
+        )
+        .map((andel) => samværsandelTilVerdi[andel])
+        .reduce((acc, andel) => acc + andel, 0);
+
+    const maksimalSamværsandel = samværsuker.length * 7 * 8;
+
+    const antallHeleDagerMedSamvær = Math.floor(summertSamvær / 8);
+
+    const rest = summertSamvær % 8;
+    const restSuffix = rest === 0 ? '' : '/8';
+
+    const prosentandel = summertSamvær / maksimalSamværsandel;
+
+    const visningstekstAntallDager = `${antallHeleDagerMedSamvær} dager og ${rest}${restSuffix} deler`;
+    const visningstekstProsentandel = `${Math.round(prosentandel * 1000) / 10}%`;
+    return `${visningstekstAntallDager} av totalt ${samværsuker.length} uker = ${visningstekstProsentandel}`;
 };
